requirements.txt
pandas
pyyaml
python-dotenv
numpy
torch
ase
tqdm
pydantic
langchain-core
langchain-openai
langchain-anthropic
langchain-deepseek
matplotlib
scipy
scikit-learn
pymatgen
lmdb
submitit
hydra-core
omegaconf
e3nn
torch-geometric
torch-scatter
